frontend: refactor TaskType -> Task
