+++ frontend/src/components/TaskForm.tsx
@@ -3,8 +3,8 @@ import { Button, TextField } from "@mui/material";
 import { Container, Stack } from "@mui/system";
 import { useForm } from "react-hook-form";
 
-import TaskContext from "../contexts/task-ocntext";
-import { TaskType } from "./taskTypes";
+import TaskContext from "../contexts/task-context";
+import { Task } from "../types/task";
 import client from "../utils/api";
 import AuthContext from "../contexts/auth-context";
 
@@ -26,7 +26,7 @@ const TaskForm: React.FC = () => {
   };
 
   // Handler for Add Task button.
-  const addTaskHandler = (task: TaskType) => {
+  const addTaskHandler = (task: Task) => {
     // send POST request and append the response data to array `tasks`.
     // The response comes with an ID generated by the backend DB.
     client.post("/tasks", task, authHead).then((response) => {
