+++ frontend/src/index.tsx
@@ -5,7 +5,7 @@ import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { AuthContextProvider } from "./contexts/auth-context";
 import { BrowserRouter } from "react-router-dom";
-import { TaskContextProvider } from "./contexts/task-ocntext";
+import { TaskContextProvider } from "./contexts/task-context";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
